moving time-format to core and reusing it
