+++ web/libs/core/src/lib/utils/helpers.ts
@@ -47,3 +47,39 @@ export const formatFileSize = (bytes: number): string => {
   const i = Math.floor(Math.log(bytes) / Math.log(k));
   return `${Number.parseFloat((bytes / k ** i).toFixed(2))} ${sizes[i]}`;
 };
+
+/**
+ * Formats seconds into a human-readable time string.
+ * Format: [n]h [n]m [n]s (omitting zero-value components)
+ * Examples:
+ *   36922 -> "10h 15m 22s"
+ *   322   -> "5m 22s"
+ *   45    -> "45s"
+ *   3600  -> "1h"
+ *   0     -> ""
+ */
+export const formatTime = (totalSeconds: number): string => {
+  const seconds = Math.floor(totalSeconds);
+
+  if (seconds <= 0) {
+    return "";
+  }
+
+  const hours = Math.floor(seconds / 3600);
+  const minutes = Math.floor((seconds % 3600) / 60);
+  const secs = seconds % 60;
+
+  const parts: string[] = [];
+
+  if (hours > 0) {
+    parts.push(`${hours}h`);
+  }
+  if (minutes > 0) {
+    parts.push(`${minutes}m`);
+  }
+  if (secs > 0) {
+    parts.push(`${secs}s`);
+  }
+
+  return parts.join(" ");
+};
